requirements.txt
Django==1.7.1
django-debug-toolbar==1.2.2
django-extensions==1.4.9
django-filter==0.10.0
djangorestframework==3.1.3
google-api-python-client==1.3.1
httplib2==0.9
Markdown==2.6.2
MySQL-python==1.2.5
oauth2client==1.4.2
pyasn1==0.1.7
pyasn1-modules==0.0.5
python-dateutil==2.3
pytz==2014.10
rsa==3.1.4
simplejson==3.6.5
six==1.8.0
sqlparse==0.1.14
uritemplate==0.6
